model_loader moved to tools
